pyproject.toml
[project]
name = "ytdlbot"
version = "1.0.0"
description = "Default template for PDM package"
authors = [
    {name = "Benny", email = "benny.think@gmail.com"},
]
dependencies = ["tgcrypto>=1.2.5", "yt-dlp[curl-cffi,default]==2025.11.12", "APScheduler>=3.11.0", "ffmpeg-python>=0.2.0", "PyMySQL>=1.1.1", "filetype>=1.2.0", "beautifulsoup4>=4.12.3", "fakeredis>=2.26.2", "redis==6.4.0", "requests>=2.32.3", "tqdm>=4.67.1", "token-bucket>=0.3.0", "python-dotenv>=1.0.1", "black>=24.10.0", "sqlalchemy>=2.0.36", "psutil>=6.1.0", "ffpb>=0.4.1", "kurigram==2.2.15", "cryptography>=43.0.3", "yt-dlp-ejs>=0.3.1", "fastapi>=0.115.0", "uvicorn[standard]>=0.32.0"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "Apache2.0"}


[tool.pdm]
distribution = false
